pyproject.toml
[project]
name = "github-tg-bot"
version = "0.1.0"
description = "Telegram bot for GitHub repository notifications"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.0",
    "python-dotenv>=1.2.0",
    "aiofiles>=24.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

